refactor(ui): move env-editor files into subdirectory, jsonValidation into editor
