+++ src/ui/AppInner.tsx
@@ -3,7 +3,7 @@ import type { Dispatch, SetStateAction } from "react"
 import { resolve } from "node:path"
 import { useKeymap } from "@opentui/keymap/react"
 import { MainView } from "./MainView"
-import { EnvironmentEditorView } from "./EnvironmentEditorView"
+import { EnvironmentEditorView } from "./env-editor/EnvironmentEditorView"
 import { AppOverlays } from "./AppOverlays"
 import { useCollection } from "../hooks/useCollection"
 import { useTreeNavigation } from "../hooks/useTreeNavigation"
@@ -25,7 +25,7 @@ import { buildCommandPaletteCommands } from "./commands"
 import { useTheme } from "./theme"
 import { StatusBar } from "./StatusBar"
 import { showToast } from "./Toast"
-import { type EnvHeaderPaneHandle } from "./EnvHeaderPane"
+import { type EnvHeaderPaneHandle } from "./env-editor/EnvHeaderPane"
 
 import { type Keybinds, displayKey } from "./keybind"
 import { useSaveFile } from "./useSaveFile"
