+++ src/ui/RequestPane.tsx
@@ -27,7 +27,7 @@ import { Select, type SelectItem } from "./Select"
 import { FormEditor } from "./FormEditor"
 import { ValidationNotice } from "./editor/ValidationNotice"
 import type { BodyType } from "../schema"
-import { validateJsonContent } from "./jsonValidation"
+import { validateJsonContent } from "./editor/jsonValidation"
 import { getEnvVarHighlights } from "./variable-completion/variableCompletion"
 
 interface Props {
