+++ src/ui/useOverlayIntercepts.ts
@@ -4,7 +4,7 @@ import { useKeymap } from "@opentui/keymap/react"
 import type { SaveState } from "./saveState"
 import type { Focus } from "./focus"
 import type { UseEnvironmentEditorResult } from "../hooks/useEnvironmentEditor"
-import type { EnvHeaderPaneHandle } from "./EnvHeaderPane"
+import type { EnvHeaderPaneHandle } from "./env-editor/EnvHeaderPane"
 import type { NewRequestOverlayHandle } from "./overlays/NewRequestOverlay"
 import type { CloneRequestOverlayHandle } from "./overlays/CloneRequestOverlay"
 import type { NewFolderOverlayHandle } from "./overlays/NewFolderOverlay"
